requirements.txt
python-telegram-bot==20.0b0
azure-cognitiveservices-vision-computervision==0.7.0
msrest==0.6.21
